Trait apple_codesign::cryptography::PrivateKey

pub trait PrivateKey: KeyInfoSigner {
    // Required methods
    fn as_key_info_signer(&self) -> &dyn KeyInfoSigner;
    fn to_public_key_peer_decrypt(
        &self
    ) -> Result<Box<dyn PublicKeyPeerDecrypt>, AppleCodesignError>;
    fn finish(&self) -> Result<(), AppleCodesignError>;
}

A supertrait generically describing a private key capable of signing and possibly decryption.

Required Methods

fn as_key_info_signer(&self) -> &dyn KeyInfoSigner

fn to_public_key_peer_decrypt( &self ) -> Result<Box<dyn PublicKeyPeerDecrypt>, AppleCodesignError>

fn finish(&self) -> Result<(), AppleCodesignError>

Signals the end of operations on the private key.

Implementations can use this to do things like destroy private key matter, disconnect from a hardware device, etc.