Enum apple_codesign::KnownCertificate

pub enum KnownCertificate {

    AppleComputerIncRoot,
    AppleRootCa,
    AppleRootCaG2Root,
    AppleRootCaG3Root,
    AppleIstCa2G1,
    AppleIstCa8G1,
    ApplicationIntegration,
    ApplicationIntegration2,
    ApplicationIntegrationG3,
    AppleApplicationIntegrationCa5G1,
    DeveloperAuthentication,
    DeveloperIdG1,
    DeveloperIdG2,
    SoftwareUpdate,
    Timestamp,
    WwdrG1,
    WwdrG2,
    WwdrG3,
    WwdrG4,
    WwdrG5,
    WwdrG6,

}

Defines all known Apple certificates.

This crate embeds the raw certificate data for the various known Apple certificate authorities, as advertised at https://www.apple.com/certificateauthority/.

This enumeration defines all the ones we know about. Instances can be dereferenced into concrete CapturedX509Certificate to get at the underlying certificate and access its metadata.

Variants

AppleComputerIncRoot

Apple Computer, Inc. Root Certificate.

C = US, O = “Apple Computer, Inc.”, OU = Apple Computer Certificate Authority, CN = Apple Root Certificate Authority

AppleRootCa

Apple Inc. Root Certificate

C = US, O = Apple Inc., OU = Apple Certification Authority, CN = Apple Root CA

AppleRootCaG2Root

Apple Root CA - G2 Root Certificate

CN = Apple Root CA - G2, OU = Apple Certification Authority, O = Apple Inc., C = US

AppleRootCaG3Root

Apple Root CA - G3 Root Certificate

CN = Apple Root CA - G3, OU = Apple Certification Authority, O = Apple Inc., C = US

AppleIstCa2G1

Apple IST CA 2 - G1 Certificate

CN = Apple IST CA 2 - G1, OU = Certification Authority, O = Apple Inc., C = US

AppleIstCa8G1

Apple IST CA 8 - G1 Certificate

CN = Apple IST CA 8 - G1, OU = Certification Authority, O = Apple Inc., C = US

ApplicationIntegration

Application Integration Certificate

C = US, O = Apple Inc., OU = Apple Certification Authority, CN = Apple Application Integration Certification Authority

ApplicationIntegration2

Application Integration 2 Certificate

CN = Apple Application Integration 2 Certification Authority, OU = Apple Certification Authority, O = Apple Inc., C = US

ApplicationIntegrationG3

Application Integration - G3 Certificate

CN = Apple Application Integration CA - G3, OU = Apple Certification Authority, O = Apple Inc., C = US

AppleApplicationIntegrationCa5G1

Apple Application Integration CA 5 - G1 Certificate

CN = Apple Application Integration CA 5 - G1, OU = Apple Certification Authority, O = Apple Inc., C = US

DeveloperAuthentication

Developer Authentication Certificate

CN = Developer Authentication Certification Authority, OU = Apple Worldwide Developer Relations, O = Apple Inc., C = US

DeveloperIdG1

Developer ID - G1 Certificate

CN = Developer ID Certification Authority, OU = Apple Certification Authority, O = Apple Inc., C = US

DeveloperIdG2

Developer ID - G2 Certificate.

CN = Developer ID Certification Authority, OU = G2, O = Apple Inc., C = US

SoftwareUpdate

Software Update Certificate

CN = Apple Software Update Certification Authority, OU = Certification Authority, O = Apple Inc., C = US

Timestamp

Timestamp Certificate

CN = Apple Timestamp Certification Authority, OU = Apple Certification Authority, O = Apple Inc., C = US

WwdrG1

Worldwide Developer Relations - G1 (Expiring 02/07/2023 21:48:47 UTC) Certificate

C = US, O = Apple Inc., OU = Apple Worldwide Developer Relations, CN = Apple Worldwide Developer Relations Certification Authority

WwdrG2

Worldwide Developer Relations - G2 (Expiring 05/06/2029 23:43:24 UTC) Certificate

CN = Apple Worldwide Developer Relations CA - G2, OU = Apple Certification Authority, O = Apple Inc., C = US

WwdrG3

Worldwide Developer Relations - G3 (Expiring 02/20/2030 00:00:00 UTC) Certificate

CN = Apple Worldwide Developer Relations Certification Authority, OU = G3, O = Apple Inc., C = US

WwdrG4

Worldwide Developer Relations - G4 (Expiring 12/10/2030 00:00:00 UTC) Certificate

CN = Apple Worldwide Developer Relations Certification Authority, OU = G4, O = Apple Inc., C = US

WwdrG5

Worldwide Developer Relations - G5 (Expiring 12/10/2030 00:00:00 UTC) Certificate

CN = Apple Worldwide Developer Relations Certification Authority, OU = G5, O = Apple Inc., C = US

WwdrG6

Worldwide Developer Relations - G6 (Expiring 03/19/2036 00:00:00 UTC) Certificate

CN = Apple Worldwide Developer Relations Certification Authority, OU = G6, O = Apple Inc., C = US

Implementations

impl KnownCertificate

pub fn all() -> &'static [&'static CapturedX509Certificate]

Obtain a slice of all known KnownCertificate.

If you want to iterate over all certificates and find one, you can use this.

pub fn all_roots() -> &'static [&'static CapturedX509Certificate]

All of Apple’s known root certificate authority certificates.