Enum apple_codesign::SettingsScope

pub enum SettingsScope {
    Main,
    Path(String),
    MultiArchIndex(usize),
    MultiArchCpuType(CpuType),
    PathMultiArchIndex(String, usize),
    PathMultiArchCpuType(String, CpuType),
}

Denotes the scope for a setting.

Settings have an associated scope defined by this type. This allows settings to apply to exactly what you want them to apply to.

Scopes can be converted from a string representation. The following syntax is recognized:

@main - Maps to SettingsScope::Main

@<int> - e.g. @0. Maps to SettingsScope::MultiArchIndex.Index

@[cpu_type=<int>] - e.g. @[cpu_type=7]. Maps to SettingsScope::MultiArchCpuType.

@[cpu_type=<string>] - e.g. @[cpu_type=x86_64]. Maps to SettingsScope::MultiArchCpuType for recognized string values (see below).

<string> - e.g. path/to/file. Maps to SettingsScope::Path.

<string>@<int> - e.g. path/to/file@0. Maps to SettingsScope::PathMultiArchIndex.

<string>@[cpu_type=<int>] - e.g. path/to/file@[cpu_type=7]. Maps to SettingsScope::PathMultiArchCpuType.

<string>@[cpu_type=<string>] - e.g. path/to/file@[cpu_type=arm64]. Maps to SettingsScope::PathMultiArchCpuType for recognized string values (see below).

Recognized cpu_type String Values

The following cpu_type= string values are recognized:

arm -> CPU_TYPE_ARM

arm64 -> CPU_TYPE_ARM64

arm64_32 -> CPU_TYPE_ARM64_32

x86_64 -> CPU_TYPE_X86_64

Variants

Main

The main entity being signed.

Can be a Mach-O file, a bundle, or any other primitive this crate supports signing.

When signing a bundle or any primitive with nested elements (such as a fat/universal Mach-O binary), settings can propagate to nested elements.

Path(String)

Filesystem path.

Can refer to a Mach-O file, a nested bundle, or any other filesystem based primitive that can be traversed into when performing nested signing.

The string value refers to the filesystem relative path of the entity relative to the main entity being signed.

MultiArchIndex(usize)

A single Mach-O binary within a fat/universal Mach-O binary.

The binary to operate on is defined by its 0-based index within the fat/universal Mach-O container.

MultiArchCpuType(CpuType)

A single Mach-O binary within a fat/universal Mach-O binary.

The binary to operate on is defined by its CPU architecture.

PathMultiArchIndex(String, usize)

Combination of SettingsScope::Path and SettingsScope::MultiArchIndex.

This refers to a single Mach-O binary within a fat/universal binary at a given relative path.

PathMultiArchCpuType(String, CpuType)

Combination of SettingsScope::Path and SettingsScope::MultiArchCpuType.

This refers to a single Mach-O binary within a fat/universal binary at a given relative path.