Function apple_codesign::create_self_signed_code_signing_certificate

pub fn create_self_signed_code_signing_certificate(
    algorithm: KeyAlgorithm,
    profile: CertificateProfile,
    team_id: &str,
    person_name: &str,
    country: &str,
    validity_duration: Duration
) -> Result<(CapturedX509Certificate, InMemorySigningKeyPair, Document), AppleCodesignError>

Create a new self-signed X.509 certificate suitable for signing code.

The created certificate contains all the extensions needed to convey that it is used for code signing and should resemble certificates.

However, because the certificate isn’t signed by Apple or another trusted certificate authority, binaries signed with the certificate may not pass Apple’s verification requirements and the OS may refuse to proceed. Needless to say, only use certificates generated with this function for testing purposes only.