Module apple_codesign::notarization

Apple notarization functionality.

Notarization works by uploading a payload to Apple servers and waiting for Apple to scan the submitted content. If Apple is appeased by your submission, they issue a notarization ticket, which can be downloaded and stapled (just a fancy word for attached) to the content you upload.

This module implements functionality for uploading content to Apple and waiting on the availability of a notarization ticket.

Modules

notary_api

App Store Connect Notary API.

Structs

Notarizer

An entity for performing notarizations.

Enums

NotarizationUpload

Represents the result of a notarization upload.

Functions

bundle_to_zip

Produce zip file data from a DirectoryBundle.