Module apple_codesign::ticket_lookup

Support for retrieving notarization tickets and stapling artifacts.

Structs

Field

TicketLookupRequest

Main JSON request object for ticket lookup requests.

TicketLookupRequestRecord

Represents a single record to look up in a ticket lookup request.

TicketLookupResponse

Main JSON response object to ticket lookup requests.

TicketLookupResponseRecordFailure

TicketLookupResponseRecordSuccess

Represents a successful ticket lookup response record.

TicketRecordEvent

Represents an event in a ticket record.

Enums

TicketLookupResponseRecord

Describes the results of a ticket lookup for a specific record.

Constants

APPLE_TICKET_LOOKUP_URL

URL of HTTP service where Apple publishes stapling tickets.

Functions

default_client

Obtain the default Client to use for HTTP requests.

lookup_notarization_ticket

Look up a single notarization ticket.

lookup_notarization_tickets

Look up a notarization ticket given an HTTP client and an iterable of record names.